# ibs: don't pas slot as pointer