--- a/core/state/cached_reader.go
+++ b/core/state/cached_reader.go
@@ -59,7 +59,7 @@ func (cr *CachedReader) ReadAccountDataForDebug(address common.Address) (*accoun
 }
 
 // ReadAccountStorage is called when a storage item needs to be fetched from the state
-func (cr *CachedReader) ReadAccountStorage(address common.Address, key *common.Hash) ([]byte, error) {
+func (cr *CachedReader) ReadAccountStorage(address common.Address, key common.Hash) ([]byte, error) {
 	addrBytes := address.Bytes()
 	if s, ok := cr.cache.GetStorage(addrBytes, 1, key.Bytes()); ok {
 		return s, nil