--- a/core/state/database.go
+++ b/core/state/database.go
@@ -37,7 +37,7 @@ const (
 type StateReader interface {
 	ReadAccountData(address common.Address) (*accounts.Account, error)
 	ReadAccountDataForDebug(address common.Address) (*accounts.Account, error)
-	ReadAccountStorage(address common.Address, key *common.Hash) ([]byte, error)
+	ReadAccountStorage(address common.Address, key common.Hash) ([]byte, error)
 	ReadAccountCode(address common.Address) ([]byte, error)
 	ReadAccountCodeSize(address common.Address) (int, error)
 	ReadAccountIncarnation(address common.Address) (uint64, error)