--- a/core/state/database_test.go
+++ b/core/state/database_test.go
@@ -177,7 +177,7 @@ func TestCreate2Revive(t *testing.T) {
 		}
 		// We expect number 0x42 in the position [2], because it is the block number 2
 		key2 = common.BigToHash(big.NewInt(2))
-		st.GetState(create2address, &key2, &check2)
+		st.GetState(create2address, key2, &check2)
 		if check2.Uint64() != 0x42 {
 			t.Errorf("expected 0x42 in position 2, got: %x", check2.Uint64())
 		}
@@ -214,12 +214,12 @@ func TestCreate2Revive(t *testing.T) {
 		// We expect number 0x42 in the position [4], because it is the block number 4
 		key4 := common.BigToHash(big.NewInt(4))
 		var check4 uint256.Int
-		st.GetState(create2address, &key4, &check4)
+		st.GetState(create2address, key4, &check4)
 		if check4.Uint64() != 0x42 {
 			t.Errorf("expected 0x42 in position 4, got: %x", check4.Uint64())
 		}
 		// We expect number 0x0 in the position [2], because it is the block number 4
-		st.GetState(create2address, &key2, &check2)
+		st.GetState(create2address, key2, &check2)
 		if !check2.IsZero() {
 			t.Errorf("expected 0x0 in position 2, got: %x", check2)
 		}
@@ -613,7 +613,7 @@ func TestReorgOverSelfDestruct(t *testing.T) {
 		}
 
 		// Remember value of field "x" (storage item 0) after the first block, to check after rewinding
-		st.GetState(contractAddress, &key0, &correctValueX)
+		st.GetState(contractAddress, key0, &correctValueX)
 		return nil
 	})
 	require.NoError(t, err)
@@ -645,7 +645,7 @@ func TestReorgOverSelfDestruct(t *testing.T) {
 			t.Error("expected contractAddress to exist at the block 4", contractAddress.String())
 		}
 		var valueX uint256.Int
-		st.GetState(contractAddress, &key0, &valueX)
+		st.GetState(contractAddress, key0, &valueX)
 		if valueX != correctValueX {
 			t.Fatalf("storage value has changed after reorg: %x, expected %x", valueX, correctValueX)
 		}
@@ -766,7 +766,7 @@ func TestReorgOverStateChange(t *testing.T) {
 		}
 
 		// Remember value of field "x" (storage item 0) after the first block, to check after rewinding
-		st.GetState(contractAddress, &key0, &correctValueX)
+		st.GetState(contractAddress, key0, &correctValueX)
 		return nil
 	})
 	require.NoError(t, err)
@@ -790,7 +790,7 @@ func TestReorgOverStateChange(t *testing.T) {
 
 		// Reload blockchain from the database
 		var valueX uint256.Int
-		st.GetState(contractAddress, &key0, &valueX)
+		st.GetState(contractAddress, key0, &valueX)
 		if valueX != correctValueX {
 			t.Fatalf("storage value has changed after reorg: %x, expected %x", valueX, correctValueX)
 		}
@@ -901,7 +901,7 @@ func TestCreateOnExistingStorage(t *testing.T) {
 			t.Error("expected contractAddress to exist at the block 1", contractAddress.String())
 		}
 
-		st.GetState(contractAddress, &key0, &check0)
+		st.GetState(contractAddress, key0, &check0)
 		if !check0.IsZero() {
 			t.Errorf("expected 0x00 in position 0, got: %x", check0.Bytes())
 		}
@@ -942,20 +942,20 @@ func TestReproduceCrash(t *testing.T) {
 		t.Errorf("error finalising 1st tx: %v", err)
 	}
 	// Start the 2nd transaction
-	intraBlockState.SetState(contract, &storageKey1, *value1)
+	intraBlockState.SetState(contract, storageKey1, *value1)
 	if err := intraBlockState.FinalizeTx(&chain.Rules{}, tsw); err != nil {
 		t.Errorf("error finalising 1st tx: %v", err)
 	}
 	// Start the 3rd transaction
 	intraBlockState.AddBalance(contract, uint256.NewInt(1000000000), tracing.BalanceChangeUnspecified)
-	intraBlockState.SetState(contract, &storageKey2, *value2)
+	intraBlockState.SetState(contract, storageKey2, *value2)
 	if err := intraBlockState.FinalizeTx(&chain.Rules{}, tsw); err != nil {
 		t.Errorf("error finalising 1st tx: %v", err)
 	}
 	// Start the 4th transaction - clearing both storage cells
 	intraBlockState.SubBalance(contract, uint256.NewInt(1000000000), tracing.BalanceChangeUnspecified)
-	intraBlockState.SetState(contract, &storageKey1, *value0)
-	intraBlockState.SetState(contract, &storageKey2, *value0)
+	intraBlockState.SetState(contract, storageKey1, *value0)
+	intraBlockState.SetState(contract, storageKey2, *value0)
 	if err := intraBlockState.FinalizeTx(&chain.Rules{}, tsw); err != nil {
 		t.Errorf("error finalising 1st tx: %v", err)
 	}
@@ -1639,7 +1639,7 @@ func TestRecreateAndRewind(t *testing.T) {
 			t.Errorf("expected phoenix %x to exist after first insert", phoenixAddress)
 		}
 
-		st.GetState(phoenixAddress, &key0, &check0)
+		st.GetState(phoenixAddress, key0, &check0)
 		if check0.Cmp(uint256.NewInt(2)) != 0 {
 			t.Errorf("expected 0x02 in position 0, got: 0x%x", check0.Bytes())
 		}
@@ -1660,7 +1660,7 @@ func TestRecreateAndRewind(t *testing.T) {
 			t.Errorf("expected phoenix %x to exist after second insert", phoenixAddress)
 		}
 
-		st.GetState(phoenixAddress, &key0, &check0)
+		st.GetState(phoenixAddress, key0, &check0)
 		if check0.Cmp(uint256.NewInt(1)) != 0 {
 			t.Errorf("expected 0x01 in position 0, got: 0x%x", check0.Bytes())
 		}
@@ -1680,7 +1680,7 @@ func TestRecreateAndRewind(t *testing.T) {
 			t.Errorf("expected phoenix %x to exist after second insert", phoenixAddress)
 		}
 
-		st.GetState(phoenixAddress, &key0, &check0)
+		st.GetState(phoenixAddress, key0, &check0)
 		if check0.Cmp(uint256.NewInt(0)) != 0 {
 			t.Errorf("expected 0x00 in position 0, got: 0x%x", check0.Bytes())
 		}