--- a/core/tracing/hooks.go
+++ b/core/tracing/hooks.go
@@ -47,7 +47,7 @@ type IntraBlockState interface {
 	GetBalance(common.Address) (*uint256.Int, error)
 	GetNonce(common.Address) (uint64, error)
 	GetCode(common.Address) ([]byte, error)
-	GetState(addr common.Address, key *common.Hash, value *uint256.Int) error
+	GetState(addr common.Address, key common.Hash, value *uint256.Int) error
 	Exist(common.Address) (bool, error)
 	GetRefund() uint64
 }
@@ -155,7 +155,7 @@ type (
 	CodeChangeHook = func(addr common.Address, prevCodeHash common.Hash, prevCode []byte, codeHash common.Hash, code []byte)
 
 	// StorageChangeHook is called when the storage of an account changes.
-	StorageChangeHook = func(addr common.Address, slot *common.Hash, prev, new uint256.Int)
+	StorageChangeHook = func(addr common.Address, slot common.Hash, prev, new uint256.Int)
 
 	// LogHook is called when a log is emitted.
 	LogHook = func(log *types.Log)