--- a/core/vm/operations_acl.go
+++ b/core/vm/operations_acl.go
@@ -45,7 +45,7 @@ func makeGasSStoreFunc(clearingRefund uint64) gasFunc {
 			cost    = uint64(0)
 		)
 
-		evm.IntraBlockState().GetState(contract.Address(), &slot, &current)
+		evm.IntraBlockState().GetState(contract.Address(), slot, &current)
 		// If the caller cannot afford the cost, this change will be rolled back
 		if _, slotMod := evm.IntraBlockState().AddSlotToAccessList(contract.Address(), slot); slotMod {
 			cost = params.ColdSloadCostEIP2929
@@ -60,7 +60,7 @@ func makeGasSStoreFunc(clearingRefund uint64) gasFunc {
 		}
 		var original uint256.Int
 		slotCommited := common.Hash(x.Bytes32())
-		evm.IntraBlockState().GetCommittedState(contract.Address(), &slotCommited, &original)
+		evm.IntraBlockState().GetCommittedState(contract.Address(), slotCommited, &original)
 		if original.Eq(&current) {
 			if original.IsZero() { // create slot (2.1.1)
 				return cost + params.SstoreSetGasEIP2200, nil