--- a/core/vm/program/program.go
+++ b/core/vm/program/program.go
@@ -26,7 +26,7 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/ethereum/go-ethereum/core/vm"
+	"github.com/erigontech/erigon/core/vm"
 	"github.com/holiman/uint256"
 )