it was some Erigon1 optimization
```
cd core/vm/runtime
go test -bench=BenchmarkSimple -run=Benchmark -count 10 . > old.txt
go test -bench=BenchmarkSimple -run=Benchmark -count 10 . > new.txt
benchstat old.txt new.txt

goos: linux
goarch: amd64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: Intel(R) Xeon(R) E-2386G CPU @ 3.50GHz
                                       │   old.txt    │              new.txt               │
                                       │    sec/op    │   sec/op     vs base               │
SimpleLoop/staticcall-identity-100M-12   323.3m ±  9%   334.1m ± 7%       ~ (p=0.165 n=10)
SimpleLoop/call-identity-100M-12         387.1m ±  4%   387.9m ± 8%       ~ (p=0.912 n=10)
SimpleLoop/loop-100M-12                  293.2m ±  4%   281.9m ± 4%  -3.84% (p=0.029 n=10)
SimpleLoop/loop2-100M-12                 358.8m ±  3%   340.6m ± 2%  -5.09% (p=0.000 n=10)
SimpleLoop/call-nonexist-100M-12         343.7m ± 11%   348.9m ± 4%       ~ (p=0.280 n=10)
SimpleLoop/call-EOA-100M-12              369.5m ±  8%   362.8m ± 7%       ~ (p=0.739 n=10)
SimpleLoop/call-reverting-100M-12        900.7m ±  9%   896.0m ± 5%       ~ (p=0.684 n=10)
geomean                                  395.2m         391.6m       -0.93%

                                       │    old.txt    │               new.txt                │
                                       │     B/op      │     B/op       vs base               │
SimpleLoop/staticcall-identity-100M-12   67.94Mi ±  0%   67.94Mi ±  0%       ~ (p=0.863 n=10)
SimpleLoop/call-identity-100M-12         66.57Mi ±  0%   66.57Mi ±  0%       ~ (p=0.171 n=10)
SimpleLoop/loop-100M-12                  2.855Ki ± 40%   1.717Ki ± 66%       ~ (p=0.170 n=10)
SimpleLoop/loop2-100M-12                 3.531Ki ± 43%   3.531Ki ± 51%       ~ (p=0.740 n=10)
SimpleLoop/call-nonexist-100M-12         22.78Mi ±  0%   22.78Mi ±  0%       ~ (p=0.554 n=10)
SimpleLoop/call-EOA-100M-12              22.78Mi ±  0%   22.78Mi ±  0%       ~ (p=1.000 n=10)
SimpleLoop/call-reverting-100M-12        201.7Mi ±  0%   201.7Mi ±  0%       ~ (p=0.393 n=10)
geomean                                  3.331Mi         3.098Mi        -7.01%

                                       │   old.txt   │              new.txt               │
                                       │  allocs/op  │  allocs/op   vs base               │
SimpleLoop/staticcall-identity-100M-12   2.740M ± 0%   2.740M ± 0%       ~ (p=0.721 n=10)
SimpleLoop/call-identity-100M-12         2.685M ± 0%   2.685M ± 0%       ~ (p=0.675 n=10)
SimpleLoop/loop-100M-12                   14.00 ± 7%    13.00 ± 8%       ~ (p=0.170 n=10)
SimpleLoop/loop2-100M-12                  14.00 ± 7%    14.00 ± 7%       ~ (p=1.000 n=10)
SimpleLoop/call-nonexist-100M-12         746.3k ± 0%   746.3k ± 0%       ~ (p=0.545 n=10)
SimpleLoop/call-EOA-100M-12              746.3k ± 0%   746.3k ± 0%       ~ (p=1.000 n=10)
SimpleLoop/call-reverting-100M-12        3.571M ± 0%   3.571M ± 0%       ~ (p=0.183 n=10)
geomean                                  60.21k        59.57k       -1.05%
```